+++ src/data/filterIconData.js
@@ -1,0 +1,28 @@
+// ProductListPage 와 BookmarkPage 에서 활용할 필터 아이콘 데이터
+export const filterIconsData = [
+  {
+    id: 0,
+    type: "All",
+    title: "전체",
+  },
+  {
+    id: 1,
+    type: "Product",
+    title: "상품",
+  },
+  {
+    id: 2,
+    type: "Category",
+    title: "카테고리",
+  },
+  {
+    id: 3,
+    type: "Exhibition",
+    title: "기획전",
+  },
+  {
+    id: 4,
+    type: "Brand",
+    title: "브랜드",
+  },
+];
