+++ src/layouts/ProductListLayout.module.css
@@ -1,0 +1,13 @@
+.filter-icons-container {
+  display: flex;
+  flex-direction: column;
+  width: 1200px;
+  margin: 20px auto;
+}
+
+.filter-icon-container {
+  display: flex;
+  justify-content: center;
+  gap: 25px;
+  margin: 30px;
+}
